refactor Twiiter api call into helper/api/twitter
