+++ server/helpers/api/index.js
@@ -1,0 +1,2 @@
+module.exports.instagram = require('./instagram')
+module.exports.twitter = require('./twitter')
